[Bug/#92]컬러칩,큐레이션 상세보기 수정
