--- weady/weady/Sources/Feature/Curation/Models/titleLineBreak.swift
@@ -1,8 +1,0 @@
-//
-//  titleLineBreak.swift
-//  weady
-//
-//  Created by 고석현 on 8/14/25.
-//
-
-import Foundation
